refactor: move debugger code into separate file
