+++ sirc-vm/sbrc-vm/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod cpu_from_bus;
 pub mod frame_reporter;
 pub mod lines;
 pub mod path;
